requirements.txt
flask==3.0.3
flask-cors==4.0.1
google-genai~=1.10.0
matplotlib~=3.10.1
numpy~=2.2.4
omegaconf~=2.3.0
openai==1.61.1
pandas==2.2.3
pydantic==2.10.6
python-dotenv~=1.0.1
requests~=2.32.3
seaborn~=0.13.2
setuptools~=68.2.0
streamlit==1.44.1
tenacity~=9.1.2
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0